requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.7
pytest==7.4.2
pytest-flask==1.2.0
stripe==6.5.0
email-validator==2.0.0
itsdangerous==2.1.2
gunicorn==21.2.0
Flask-Babel==4.0.0

# OSINT dependencies
python-whois==0.9.6
dnspython==2.7.0
tldextract==5.3.0
beautifulsoup4==4.12.2
lxml==5.3.0

# AI Chatbot dependencies
httpx==0.27.0
anthropic>=0.50.0

# MailGuard dependencies
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msal==1.25.0
apscheduler==3.10.4
imapclient==2.3.1
phonenumbers>=8.13.0